services/ai-service/requirements.txt
# ==============================================================================
# AI Service — Dependencies
# ==============================================================================

fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0

# LLM Providers
langchain==0.1.9
langchain-openai==0.0.8
langchain-anthropic==0.1.1
langchain-google-genai==0.0.11
langchain-community==0.0.24

# gRPC
grpcio==1.62.0
grpcio-tools==1.62.0
protobuf==4.25.3

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.27

# Redis
redis[hiredis]==5.0.1

# Kafka
aiokafka==0.10.0

# ChromaDB
chromadb==0.4.24

# Observability
prometheus-client==0.20.0
python-json-logger==2.0.7
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0

# Utils
pyyaml==6.0.1
httpx==0.27.0
python-multipart==0.0.9
tiktoken==0.6.0
